Run it: snapshot refresh, dead-feed handling, docs, CI
A 25-minute soak against the replay feed at 6x, plus every page driven in a
real browser, turned up two more things the tests had not.

The first was found on the Diagnostics page: every symbol drifting to
NO_SNAPSHOT and quality 0.30 while its book was provably in sync. The diff
chain keeps a READY book correct against what the venue sent, but nothing was
ever re-verifying it against an authoritative snapshot, so the age grew
without bound until the gate blocked everything. The engine now re-seeds at
half the limit, which makes the flag mean what it says: reaching the full age
means the refresh itself is failing, usually a REST endpoint unreachable while
the WebSocket still flows.

The second appeared when the replay file ran out. That refresh was firing on a
feed that had already stopped, clearing a good book to wait for a snapshot
that would never come, so a dead feed was reported as a desync we had caused
ourselves. The refresh is now conditional on the feed actually flowing, and
the reported state is the truth: DISCONNECTED when the adapter says so, STALE
when it just went quiet.

Also: the resync task is held rather than fire-and-forget, since asyncio keeps
only a weak reference and can collect one mid-flight, leaving a desynced book
with nothing coming to fix it.

The README leads with what was not verified. This build has never opened a
socket to Binance - the container refuses the host - so the adapter is written
to the documented streams and tested against their shapes, and `diagnose` is
the first thing to run on a machine that can actually reach the venue.

+++ aurum-edge-lab/aurum/clock.py
@@ -105,7 +105,7 @@ class LatencyTracker:
         if not self._values:
             return 0.0
         ordered = sorted(self._values)
-        idx = min(len(ordered) - 1, max(0, int(round(pct / 100.0 * (len(ordered) - 1)))))
+        idx = min(len(ordered) - 1, max(0, round(pct / 100.0 * (len(ordered) - 1))))
         return ordered[idx]
 
     def to_dict(self) -> dict[str, float]:
